+++ Controllers/UsersController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -25,55 +26,76 @@ namespace P4._0_backend.Controllers
         }
 
         // GET: api/Users
+        [Authorize]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Users>>> GetUsers()
         {
-            return await _context.Users.ToListAsync();
+            if (Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == "UserLevel").Value) == 1)
+            {
+                return await _context.Users.ToListAsync();
+            }
+
+            return Unauthorized();
+            
         }
 
         // GET: api/Users/5
+        [Authorize]
         [HttpGet("{id}")]
         public async Task<ActionResult<Users>> GetUsers(int id)
         {
-            var users = await _context.Users.FindAsync(id);
-
-            if (users == null)
+            if (Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == "UserID").Value) == id)
             {
-                return NotFound();
+                var users = await _context.Users.FindAsync(id);
+
+                if (users == null)
+                {
+                    return NotFound();
+                }
+
+                return users;
             }
 
-            return users;
+            return Unauthorized();
+           
         }
 
         // PUT: api/Users/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<IActionResult> PutUsers(int id, Users users)
         {
-            if (id != users.ID)
+            if (Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == "UserID").Value) == id)
             {
-                return BadRequest();
-            }
-
-            _context.Entry(users).State = EntityState.Modified;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!UsersExists(id))
+                if (id != users.ID)
                 {
-                    return NotFound();
+                    return BadRequest();
                 }
-                else
+
+                _context.Entry(users).State = EntityState.Modified;
+
+                try
                 {
-                    throw;
+                    await _context.SaveChangesAsync();
                 }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!UsersExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+
+                return NoContent();
             }
 
-            return NoContent();
+            return Unauthorized();
+            
         }
 
         // POST: api/Users
@@ -99,19 +121,26 @@ namespace P4._0_backend.Controllers
         }
 
         // DELETE: api/Users/5
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteUsers(int id)
         {
-            var users = await _context.Users.FindAsync(id);
-            if (users == null)
+            if (Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == "UserLevel").Value) == 1)
             {
-                return NotFound();
+                var users = await _context.Users.FindAsync(id);
+                if (users == null)
+                {
+                    return NotFound();
+                }
+
+                _context.Users.Remove(users);
+                await _context.SaveChangesAsync();
+
+                return NoContent();
             }
 
-            _context.Users.Remove(users);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+            return Unauthorized();
+           
         }
 
         private bool UsersExists(int id)
